cleanup after renaming
